Fix for hipchat blank screen error
I made this little script to fix an issue on one of the ubuntu work stations that was running an older version of hipchat on version 10.17. I ran into a similar issue on centOS with regards to chrome. Sandboxing I mean. It's in a .desktop rather than a script so that the end user can just double click it to run. Change paths accordingly.